+++ blabpy/seedlings/pipeline.py
@@ -20,8 +20,8 @@ from .listened_time import listen_time_stats_for_report, _get_subregion_count, _
 from .merge import create_merged, FIXME
 from .opf import export_opf_to_csv
 from .paths import get_all_opf_paths, get_all_cha_paths, get_basic_level_path, _parse_out_child_and_month, \
-    ensure_folder_exists_and_empty, _check_modality, get_seedlings_path, get_cha_path, get_opf_path, \
-    _normalize_child_month, get_lena_5min_csv_path, get_its_path, split_recording_id, get_seedlings_nouns_private_path
+    _check_modality, get_seedlings_path, get_cha_path, get_opf_path, _normalize_child_month, get_lena_5min_csv_path, \
+    get_its_path, split_recording_id, get_seedlings_nouns_private_path
 from .regions import get_processed_audio_regions as _get_processed_audio_regions, _get_amended_regions, \
     SPECIAL_CASES as AUDIO_SPECIAL_CASES, get_top3_top4_surplus_regions as _get_top3_top4_surplus_regions, \
     are_tokens_in_top3_top4_surplus
@@ -30,6 +30,7 @@ from .regions.regions import calculate_total_listened_time_ms, calculate_total_r
 from .scatter import copy_all_basic_level_files_to_subject_files
 from .. import ANONYMIZATION_DATE
 from ..its import Its, ItsNoTimeZoneInfo
+from ..utils import ensure_folder_exists_and_empty
 
 
 def export_all_opfs_to_csv(output_folder: Path, suffix='_processed'):