+++ blabpy/vihi/pipeline.py
@@ -2,6 +2,7 @@
 This module contains function that interact with files - either reading or writing them. The idea is that all the
 other function are supposed to be pure. That is not true at all at this time though.
 """
+from importlib.metadata import version
 from pathlib import Path
 from shutil import copy2
 from tempfile import TemporaryDirectory
@@ -11,11 +12,13 @@ import pandas as pd
 
 from .intervals.intervals import add_metric, make_intervals, add_annotation_intervals_to_eaf, _region_output_files, \
     select_best_intervals, _extract_interval_info, INTERVALS_FOR_ANNOTATION_COUNT, INTERVALS_EXTRA_COUNT
+from .reliability import prepare_eaf_for_reliability, NoAnnotationsError
 from ..its import Its, ItsNoTimeZoneInfo
 from .paths import get_its_path, parse_full_recording_id, get_eaf_path, get_rttm_path
 from ..utils import df_to_list_of_tuples
 from ..vtc import read_rttm, split_rttm
-from ..eaf import EafPlus
+from ..eaf import EafPlus, eaf_to_tree, get_annotation_values, find_child_annotation_ids, get_annotations_with_parents, \
+    find_single_element, tree_to_eaf
 
 
 def gather_recordings(full_recording_id, forced_timezone=None):
@@ -160,3 +163,41 @@ def distribute_all_rttm():
             print(f'\nVTC output for {rttm_path_from.parent.name} copied to\n'
                   f'{rttm_path_to}')
             print('\n')
+
+
+def create_reliability_test_file(eaf_path, output_dir=None):
+    output_dir = output_dir or Path()
+    output_eaf_path = output_dir / f'{eaf_path.stem}_for-reliability.eaf'
+    log_path = output_dir / f'{output_eaf_path.stem}.log'
+
+    # Load both as an EafPlus and as a tree
+    eaf = EafPlus(eaf_path)
+    eaf_tree = eaf_to_tree(eaf_path)
+
+    # Prepare the file for reliability tests
+    random_seed = list(map(ord, eaf_path.stem))
+    try:
+        eaf_tree_new, (sampled_code_nums, sampled_sampling_types) = prepare_eaf_for_reliability(
+            eaf_tree, eaf, random_seed=random_seed)
+    except NoAnnotationsError:
+        # Write error message to the log
+        with log_path.open('w') as f:
+            f.write(f'No annotations in {eaf_path}\n')
+        raise NoAnnotationsError(f'No annotations in {eaf_path}')
+    except Exception as e:
+        # Write error message to the log
+        with log_path.open('w') as f:
+            f.write(f'Error in {eaf_path}\n')
+            f.write(f'{e}\n')
+        raise Exception(f'Error in {eaf_path}') from e
+
+    # Save the result and the log
+
+    tree_to_eaf(eaf_tree_new, output_eaf_path)
+    with log_path.open('w') as f:
+        f.write(f'Sampled intervals (code_num values): {sampled_code_nums}\n')
+        f.write(f'Their sampling types: {sampled_sampling_types}\n')
+        f.write(f'Name of the original file: {eaf_path.stem}\n')
+        f.write(f'Blabpy version: {version("blabpy")}\n')
+
+    return output_eaf_path, log_path